refactor(mls): split mod.rs, drop Send/Sync bound, dedup freeze sync send
- `service/mod.rs` is now manifest-only. `OpenMlsService` struct +
  `new_as_creator` / `new_from_welcome` / `generate_key_package`
  constructors moved to a sibling `service/openmls.rs`.
- `MlsService` drops the `Send + Sync` supertrait bounds, matching the
  shape we already use for `DeliveryService` and `ConsensusPlugin`.
  Integrators that move services across threads add the bound at
  their storage site.
- The freeze post-merge path was inlining the same body as
  `send_conversation_sync`; calls the existing function instead.
- `core/steward_list/plugin.rs` was the one production `use super::`
  in `src/`; switched to the full crate path.
